[CV-423] Refactored code according to Sonarcloud, finance/page.tsx and in properties/page.tsx, made a useFetchData.ts to reduce duplicability
