+++ apps/web/app/properties/page.tsx
@@ -1,10 +1,10 @@
 "use client";
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 import { IoClose } from "react-icons/io5";
 import { db } from "@web/firebase";
 import { collection, query, where, getDocs } from "firebase/firestore";
-import CondoUnitAdapter from "../../../native/components/CondoUnitAdapter";
-import PropertyAdapter from "../../../native/components/PropertyAdapter";
+import { useFetchData } from '../useFetchData';
+
 import {
   DashboardNav,
   PropertyList,
@@ -14,16 +14,14 @@ import {
   FacilityBookingPage,
 } from "@ui/index";
 
-const Page = () => {
-  const [ownedProperties, setOwnedProperties] = useState<Object[]>([]);
-  const [isLoading, setIsLoading] = useState<Boolean>(true);
+const Page: React.FC = () => {
   const [isModalOpen, setIsModalOpen] = useState<Boolean>(false);
-  const [selectedProperty, setSelectedProperty] = useState<any>();
   /*** Used for creating the right amount of units and to pair them properly.*/
   const [propertyId, setPropertyId] = useState("");
   const [unitCount, setUnitCount] = useState(0);
   const [currentUnit, setCurrentUnit] = useState(0); // Start with 0 to not display
   const [showBookingModal, setShowBookingModal] = useState(false);
+  const { ownedProperties, selectedProperty, isLoading, setSelectedProperty } = useFetchData();
 
   // Function to toggle the booking modal
   const toggleBookingModal = () => {
@@ -57,76 +55,6 @@ const Page = () => {
     }
   };
 
-  useEffect(() => {
-    const fetchData = async () => {
-      const propertyList: Object[] = [];
-      try {
-        const propertiesCollectionSnapshot = await getDocs(
-          collection(db, "properties")
-        );
-
-        propertiesCollectionSnapshot.forEach(async (propertyDoc) => {
-          const unitList: Object[] = [];
-          const condoUnitsSnapshot = await getDocs(
-            collection(db, "properties", propertyDoc.id, "condoUnits")
-          );
-          const userUID = await localStorage.getItem("userUID");
-          let propertyData = propertyDoc.data();
-
-          condoUnitsSnapshot.forEach((condoUnitDoc) => {
-            const condoData = condoUnitDoc.data();
-            const condoId = condoUnitDoc.id;
-
-            const condoUnit = new CondoUnitAdapter(
-              condoId,
-              {
-                includes: condoData?.condoFees?.includes,
-                isPayed: condoData?.condoFees?.isPayed,
-                monthlyFee: condoData?.condoFees?.monthlyFee,
-              },
-              condoData.lockerId,
-              {
-                contact: condoData?.occupantInfo?.contact,
-                name: condoData?.occupantInfo?.name,
-              },
-              condoData.owner,
-              condoData.parkingSpotId,
-              condoData.size,
-              condoData.unitId
-            );
-            unitList.push(condoUnit.toJSON());
-          });
-          if (propertyData.owner === userUID) {
-            const property = new PropertyAdapter(
-              propertyDoc.id,
-              propertyData.address,
-              propertyData.lockerCount,
-              propertyData.owner,
-              propertyData.parkingCount,
-              propertyData.propertyName,
-              propertyData.unitCount,
-              propertyData.latitude,
-              propertyData.longitude,
-              unitList
-            );
-            propertyList.push(property.toJSON());
-            setOwnedProperties(propertyList);
-            if (propertyList.length > 0) {
-              setSelectedProperty(propertyList[0]);
-            }
-          }
-        });
-      } catch (error) {
-        console.error("Error fetching data:", error);
-      } finally {
-        setTimeout(() => {
-          setIsLoading(false);
-        }, 500);
-      }
-    };
-
-    fetchData();
-  }, []);
 
   return (
     <div className="flex flex-col h-screen bg-gradient-to-r from-[#87A8FA] to-[#87CCFA] overflow-hidden">
