+++ packages/ui/src/properties/PropertyListFinance.tsx
@@ -8,21 +8,25 @@ const PropertyList = ({ ownedProperties, inView, setSelectedProperty, isModalOpe
             <h1 className="text-2xl font-bold text-white pb-2 pl-2">
                 Properties
             </h1>
+            <div style={{display: 'flex', flexDirection: 'column'}}>
             {ownedProperties?.map((property: any) => {
                 return (
-                    <div key={property.id} 
-                    className={`pl-2 pr-10 py-4 border-b border-blue-500  ${selected == property.id ? 'bg-blue-500' : 'cursor-pointer'}`} 
-                    onClick={() => {
-                        setSelected(property.id);
-                        setSelectedProperty(property);
-                    }}>
+                    <button 
+                        key={property.id}
+                        className={`pl-2 pr-10 py-4 border-b border-blue-500  ${selected == property.id ? 'bg-blue-500' : 'cursor-pointer'}`}
+                        onClick={() => {
+                            setSelected(property.id);
+                            setSelectedProperty(property);
+                        }}
+                    >
                         <h1 className="font-bold text-xl capitalize">
                             {property.propertyName}
                         </h1>
                         <p>{property.address}</p>
-                    </div>
+                    </button>
                 );
             })}
+            </div>
         </div>
     );
 };
